blog_python/requirements.txt
python-dotenv==1.0.0
openai>=1.7.1,<2.0.0
aztp-client==1.0.7
python-dateutil==2.8.2
langchain>=0.1.0,<0.2.0
crewai==0.11.0
setuptools>=69.0.3 